app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.124.2",
    "jinja2>=3.1.6",
    "mysql-connector-python>=9.5.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
